rename ImagePopupHelper file
