+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/utility/image-popup-helper.ts
